pangob2d: Implemented basic font family caching
Despite Blend2D providing the number of families present
in the system, it does not provide a way to iterate
through all of them, requiring me the usage of fontconfig.

This implementation of font loading is highly inefficient,
because it loads all fonts (fonts, not families) present in
the system at once in memory. For example, my system has 798
fonts installed. This made the memory consumption jump from
25MB to 28MB. [note 1].